Refactor `AppHeaderComponent` to `ClueAppHeaderComponent`
