+++ src/clue/components/clue-app-header.sass
@@ -1,4 +1,4 @@
-@import ./vars
+@import ../../components/vars
 
 $member-size: 18px
 
@@ -126,7 +126,7 @@ $member-size: 18px
       .profile-icon-inner
         width: 30px
         height: 30px
-        mask-image: url("../assets/icons/clue-dashboard/teacher-student.svg")
+        mask-image: url("../../assets/icons/clue-dashboard/teacher-student.svg")
         background-color: $sage-light-2
 
     .user
