+++ src/clue/components/clue-app-header.tsx
@@ -1,27 +1,15 @@
-import { Button, ButtonGroup } from "@blueprintjs/core";
 import { inject, observer } from "mobx-react";
 import * as React from "react";
-import { BaseComponent, IBaseProps } from "./base";
-import { ClassMenuContainer } from "./class-menu-container";
-import { ProblemMenuContainer } from "./problem-menu-container";
+import { EPanelId, IPanelGroupSpec } from "../../components/app-header";
+import { BaseComponent, IBaseProps } from "../../components/base";
+import { ClassMenuContainer } from "../../components/class-menu-container";
+import { ProblemMenuContainer } from "../../components/problem-menu-container";
 import { ToggleGroup, Themes } from "@concord-consortium/react-components";
-import { GroupModelType, GroupUserModelType } from "../models/stores/groups";
+import { GroupModelType, GroupUserModelType } from "../../models/stores/groups";
 
-import "./utilities/blueprint.sass";
-import "./app-header.sass";
+import "../../components/utilities/blueprint.sass";
+import "./clue-app-header.sass";
 const Colors = Themes.Default;
-
-export enum EPanelId {
-  dashboard = "dashboard",
-  workspace = "workspace"
-}
-
-export interface IPanelSpec {
-  panelId: EPanelId;
-  label: string;
-  content: JSX.Element;
-}
-export type IPanelGroupSpec = IPanelSpec[];
 
 interface IProps extends IBaseProps {
   isGhostUser: boolean;
@@ -33,7 +21,7 @@ interface IProps extends IBaseProps {
 
 @inject("stores")
 @observer
-export class AppHeaderComponent extends BaseComponent<IProps, {}> {
+export class ClueAppHeaderComponent extends BaseComponent<IProps, {}> {
 
   public render() {
     const { showGroup } = this.props;
